requirements.txt
feedparser>=6.0.11
requests>=2.32.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
openai>=1.35.0
groq>=0.9.0